+++ src/components/LoadingSpinner/LoadingSpinner.tsx
@@ -14,14 +14,14 @@ const gridStyles = makeStyles({
     flexDirection: 'column',
     width: '100%',
     animation: '0.45s fade-in',
-    alignItems: 'center'
-  }
+    alignItems: 'center',
+  },
 });
 
 const typographyStyles = makeStyles({
   root: {
-    paddingTop: '10px'
-  }
+    paddingTop: '10px',
+  },
 });
 
 function LoadingSpinner({message}: LoadingSpinnerProps): JSX.Element {